src/test/java/ru/netology/stats/StatsServiceTests.java
package ru.netology.stats;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvFileSource;
import org.junit.jupiter.params.provider.CsvSource;

public class StatsServiceTests {
    @Test
    public void shouldFindTotalSumSales() {
        StatsService service = new StatsService();

        int[] array = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 180;
        int actual = service.getTotalSumSales(array);

        Assertions.assertEquals(expected, actual);
    }

    @Test
    public void shouldFindAverageSalesAmount() {
        StatsService service = new StatsService();

        int[] array = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 15;
        int actual = service.getAverageSalesAmount(array);

        Assertions.assertEquals(expected, actual);

    }

    @Test
    public void shouldFindMonthWithBiggestSale() {
        StatsService service = new StatsService();

        int[] array = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.getMonthWithBiggestSale(array);

        Assertions.assertEquals(expected, actual);

    }
}
